requirements.txt
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
mysqlclient==2.1.1
djoser==2.1.0